Implemented siem log configuration (create, destroy)
