+++ incapsula/provider.go
@@ -118,9 +118,10 @@ func Provider() *schema.Provider {
 			"incapsula_mtls_client_to_imperva_ca_certificate":                  resourceMtlsClientToImpervaCertificate(),
 			"incapsula_mtls_client_to_imperva_ca_certificate_site_association": resourceMtlsClientToImpervaCertificateSiteAssociation(),
 			"incapsula_mtls_client_to_imperva_ca_certificate_site_settings":    resourceMtlsClientToImpervaCertificateSetings(),
-			"incapsula_s3_siem_connection":                                     resourceS3SiemConnection(),
-			"incapsula_s3arn_siem_connection":                                  resourceS3ArnSiemConnection(),
-			"incapsula_siem_log_configuration":                                 resourceSiemLogConfiguration(),
+			"incapsula_siem_connection_s3":                                     resourceSiemConnectionS3(),
+			"incapsula_siem_connection_s3arn":                                  resourceSiemConnectionS3Arn(),
+			"incapsula_siem_log_configuration_abp":                             resourceSiemLogConfigurationAbp(),
+			"incapsula_siem_log_configuration_netsec":                          resourceSiemLogConfigurationNetsec(),
 		},
 	}
