+++ incapsula/resource_siem_connnection_s3.go
@@ -1,17 +1,19 @@
 package incapsula
 
 import (
+	"crypto/sha256"
+	"encoding/hex"
 	"fmt"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
 
-func resourceS3ArnSiemConnection() *schema.Resource {
+func resourceSiemConnectionS3() *schema.Resource {
 	return &schema.Resource{
-		Create: resourceS3ArnSiemConnectionCreate,
-		Read:   resourceS3ArnSiemConnectionRead,
-		Update: resourceS3ArnSiemConnectionUpdate,
-		Delete: resourceS3ArnSiemConnectionDelete,
+		Create: resourceSiemConnectionS3Create,
+		Read:   resourceSiemConnectionS3Read,
+		Update: resourceSiemConnectionS3Update,
+		Delete: resourceSiemConnectionS3Delete,
 		Importer: &schema.ResourceImporter{
 			State: func(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
 				d.SetId(d.Id())
@@ -34,7 +36,35 @@ func resourceS3ArnSiemConnection() *schema.Resource {
 				Description:  "Type of the storage.",
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validation.StringInSlice([]string{"CUSTOMER_S3_ARN"}, false),
+				ValidateFunc: validation.StringInSlice([]string{"CUSTOMER_S3"}, false),
+			},
+			"access_key": {
+				Description: "Access key in AWS.",
+				Type:        schema.TypeString,
+				Optional:    true,
+				Sensitive:   true,
+				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+					expectedLen := 20
+					actualLen := len(val.(string))
+					if actualLen != expectedLen {
+						errs = append(errs, fmt.Errorf("%q length should be %d, got: %d", key, expectedLen, actualLen))
+					}
+					return
+				},
+			},
+			"secret_key": {
+				Description: "Secret key in AWS.",
+				Type:        schema.TypeString,
+				Optional:    true,
+				Sensitive:   true,
+				ValidateFunc: func(val interface{}, key string) (warns []string, errs []error) {
+					expectedLen := 40
+					actualLen := len(val.(string))
+					if actualLen != expectedLen {
+						errs = append(errs, fmt.Errorf("%q length should be %d, got: %d", key, expectedLen, actualLen))
+					}
+					return
+				},
 			},
 			"path": {
 				Description: "Store data from the specified connection under this path.",
@@ -46,18 +76,34 @@ func resourceS3ArnSiemConnection() *schema.Resource {
 				Type:        schema.TypeString,
 				Optional:    true,
 			},
+			"input_hash": {
+				Description: "inputHash",
+				Type:        schema.TypeString,
+				Optional:    true,
+				Computed:    true,
+				Sensitive:   true,
+				DiffSuppressFunc: func(k, old, new string, d *schema.ResourceData) bool {
+					newHash := createS3SiemConnectionHash(d)
+					if newHash == old {
+						return true
+					}
+					return false
+				},
+			},
 		},
 	}
 }
 
-func resourceS3ArnSiemConnectionCreate(d *schema.ResourceData, m interface{}) error {
+func resourceSiemConnectionS3Create(d *schema.ResourceData, m interface{}) error {
 	client := m.(*Client)
 	siemConnectionInfo := SiemConnectionInfo{
 		AssetID:        d.Get("account_id").(string),
 		ConnectionName: d.Get("connection_name").(string),
 		StorageType:    d.Get("storage_type").(string),
 		ConnectionInfo: ConnectionInfo{
-			Path: d.Get("path").(string),
+			AccessKey: d.Get("access_key").(string),
+			SecretKey: d.Get("secret_key").(string),
+			Path:      d.Get("path").(string),
 		},
 	}
 
@@ -71,13 +117,13 @@ func resourceS3ArnSiemConnectionCreate(d *schema.ResourceData, m interface{}) er
 
 	if (*responseStatusCode == 201) && (siemConnectionWithIdAndVersion != nil) && (len(siemConnectionWithIdAndVersion.Data) == 1) {
 		d.SetId(siemConnectionWithIdAndVersion.Data[0].ID)
-		return resourceS3SiemConnectionRead(d, m)
+		return resourceSiemConnectionS3Read(d, m)
 	} else {
 		return fmt.Errorf("[ERROR] Unsupported operation. Response status code: %d", *responseStatusCode)
 	}
 }
 
-func resourceS3ArnSiemConnectionRead(d *schema.ResourceData, m interface{}) error {
+func resourceSiemConnectionS3Read(d *schema.ResourceData, m interface{}) error {
 	client := m.(*Client)
 	siemConnectionWithIdAndVersion, responseStatusCode, err := client.ReadSiemConnection(d.Id())
 	if err != nil {
@@ -92,15 +138,17 @@ func resourceS3ArnSiemConnectionRead(d *schema.ResourceData, m interface{}) erro
 		d.Set("account_id", connection.AssetID)
 		d.Set("connection_name", connection.ConnectionName)
 		d.Set("storage_type", connection.StorageType)
+		d.Set("access_key", connection.ConnectionInfo.AccessKey)
 		d.Set("path", connection.ConnectionInfo.Path)
 		d.Set("version", connection.Version)
+		d.Set("input_hash", connection.ConnectionInfo.SecretKey)
 		return nil
 	} else {
 		return fmt.Errorf("[ERROR] Unsupported operation. Response status code: %d", *responseStatusCode)
 	}
 }
 
-func resourceS3ArnSiemConnectionUpdate(d *schema.ResourceData, m interface{}) error {
+func resourceSiemConnectionS3Update(d *schema.ResourceData, m interface{}) error {
 	client := m.(*Client)
 	siemConnectionWithIdAndVersionInfo := SiemConnectionWithIdAndVersionInfo{
 		ID:             d.Id(),
@@ -109,7 +157,9 @@ func resourceS3ArnSiemConnectionUpdate(d *schema.ResourceData, m interface{}) er
 		Version:        d.Get("version").(string),
 		StorageType:    d.Get("storage_type").(string),
 		ConnectionInfo: ConnectionInfo{
-			Path: d.Get("path").(string),
+			AccessKey: d.Get("access_key").(string),
+			SecretKey: d.Get("secret_key").(string),
+			Path:      d.Get("path").(string),
 		},
 	}
 
@@ -125,7 +175,7 @@ func resourceS3ArnSiemConnectionUpdate(d *schema.ResourceData, m interface{}) er
 	return nil
 }
 
-func resourceS3ArnSiemConnectionDelete(d *schema.ResourceData, m interface{}) error {
+func resourceSiemConnectionS3Delete(d *schema.ResourceData, m interface{}) error {
 	client := m.(*Client)
 	ID := d.Id()
 
@@ -139,4 +189,19 @@ func resourceS3ArnSiemConnectionDelete(d *schema.ResourceData, m interface{}) er
 	// Implicitly clears the resource
 	d.SetId("")
 	return nil
+}
+
+func createS3SiemConnectionHash(d *schema.ResourceData) string {
+	secretKey := d.Get("secret_key").(string)
+	result := calculateS3SiemConnectionHash(secretKey)
+	return result
+}
+
+func calculateS3SiemConnectionHash(secretKey string) string {
+	h := sha256.New()
+	stringForHash := secretKey
+	h.Write([]byte(stringForHash))
+	byteString := h.Sum(nil)
+	result := hex.EncodeToString(byteString)
+	return result
 }
