+++ incapsula/resource_siem_log_configuration.go
@@ -3,71 +3,15 @@ package incapsula
 import (
 	"fmt"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
-
-func resourceSiemLogConfiguration() *schema.Resource {
-	return &schema.Resource{
-		Create: resourceSiemLogConfigurationCreate,
-		Read:   resourceSiemLogConfigurationRead,
-		Update: resourceSiemLogConfigurationUpdate,
-		Delete: resourceSiemLogConfigurationDelete,
-		Importer: &schema.ResourceImporter{
-			State: func(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
-				d.SetId(d.Id())
-				return []*schema.ResourceData{d}, nil
-			},
-		},
-
-		Schema: map[string]*schema.Schema{
-			"account_id": {
-				Description: "Client account id.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"configuration_name": {
-				Description: "Name of the configuration.",
-				Type:        schema.TypeString,
-				Required:    true,
-			},
-			"provider": {
-				Description:  "Type of the provider.",
-				Type:         schema.TypeString,
-				Required:     true,
-				ValidateFunc: validation.StringInSlice([]string{"ABP", "NETSEC"}, false),
-			},
-			"datasets": {
-				Description:  "All datasets for the supported providers.",
-				Type:         schema.TypeString,
-				Required:     true,
-				ValidateFunc: validation.StringInSlice([]string{"ABP", "ABP_ACCESS", "CONNECTION", "NETFLOW, IP", "ATTACK", "ABP_TEST"}, false),
-			},
-			"enabled": {
-				Description: "True if the connection is enabled, false otherwise.",
-				Type:        schema.TypeBool,
-				Required:    true,
-			},
-			"connection_id": {
-				Description: "The id of the connection for this log configuration.",
-				Type:        schema.TypeString,
-				Required:    true,
-			},
-			"version": {
-				Description: "Version of the log configuration.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-		},
-	}
-}
 
 func resourceSiemLogConfigurationCreate(d *schema.ResourceData, m interface{}) error {
 	client := m.(*Client)
 	siemLogConfigurationInfo := SiemLogConfigurationInfo{
 		AssetID:           d.Get("account_id").(string),
 		ConfigurationName: d.Get("configuration_name").(string),
-		Provider:          d.Get("provider").(string),
-		Datasets:          d.Get("datasets").([]string),
+		Provider:          d.Get("producer").(string),
+		Datasets:          d.Get("datasets").([]interface{}),
 		Enabled:           d.Get("enabled").(bool),
 		ConnectionId:      d.Get("connection_id").(string),
 	}
@@ -100,13 +44,17 @@ func resourceSiemLogConfigurationRead(d *schema.ResourceData, m interface{}) err
 		return nil
 	} else if (*responseStatusCode == 200) && (siemLogConfigurationWithIdAndVersion != nil) && (len(siemLogConfigurationWithIdAndVersion.Data) == 1) {
 		var logConfiguration = siemLogConfigurationWithIdAndVersion.Data[0]
-		d.Set("account_id", logConfiguration.AssetID)
+		if len(logConfiguration.AssetID) > 0 {
+			d.Set("account_id", logConfiguration.AssetID)
+		}
 		d.Set("configuration_name", logConfiguration.ConfigurationName)
-		d.Set("provider", logConfiguration.Provider)
+		d.Set("producer", logConfiguration.Provider)
 		d.Set("datasets", logConfiguration.Datasets)
 		d.Set("enabled", logConfiguration.Enabled)
 		d.Set("connection_id", logConfiguration.ConnectionId)
-		d.Set("version", logConfiguration.Version)
+		if len(logConfiguration.Version) > 0 {
+			d.Set("version", logConfiguration.Version)
+		}
 		return nil
 	} else {
 		return fmt.Errorf("[ERROR] Unsupported operation. Response status code: %d", *responseStatusCode)
@@ -119,8 +67,8 @@ func resourceSiemLogConfigurationUpdate(d *schema.ResourceData, m interface{}) e
 		ID:                d.Id(),
 		AssetID:           d.Get("account_id").(string),
 		ConfigurationName: d.Get("configuration_name").(string),
-		Provider:          d.Get("provider").(string),
-		Datasets:          d.Get("datasets").([]string),
+		Provider:          d.Get("producer").(string),
+		Datasets:          d.Get("datasets").([]interface{}),
 		Enabled:           d.Get("enabled").(bool),
 		ConnectionId:      d.Get("connection_id").(string),
 		Version:           d.Get("version").(string),
